packages/substances-service/substances/presentation/rest/pyproject.toml
[tool.poetry]
name = "substances-rest"
version = "0.1.0"
description = ""
authors = ["Luidgi Gromat <gromat.luidgi@gmail.com>"]
readme = "README.md"
packages = [{include = "substances_rest"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.2"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
substances-core = { path= "../../core", develop = true}
substances-postgresql = { path= "../../database/postgresql", develop = true}
substances-etl = { path= "../../infrastructure/etl", develop = true}
substances-importers = { path= "../../infrastructure/importers", develop = true}
substances-rabbitmq = { path= "../../infrastructure/rabbitmq", develop = true}
shared-kernel = "^0.1.0"
science-kernel = "^0.1.0"
fastapi = "^0.88.0"
asyncio = "^3.4.3"
aio-pika = "^8.2.5"
jsonpickle = "^3.0.0"
sse-starlette = "^1.2.1"
graphene = "^3.2.1"
graphene-pydantic = "^0.4.0"
starlette-graphene3 = "^0.6.0"
grpcio = "^1.51.1"
protobuf = "^4.21.12"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
black = "^22.12.0"
pylint = "^2.15.8"
bandit = "^1.7.4"
flask = "^2.2.2"
isort = "^5.11.3"
flake8 = "^6.0.0"
grpcio-tools = "^1.51.1"
pylint-protobuf = "^0.20.2"
grpc-stubs = "^1.24.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "run:start"